refactor(agent): split layout into types, solver and edge assembly
layout.ts was at 389 of the 399-line limit, which left no room for the
routing work the module actually needs. Split along the seams that were
already there:

  layoutTypes.ts   the geometry vocabulary
  layoutSolver.ts  the dagre adapter, and the owner of the job/result types
  layoutEdges.ts   arrow points and label positions
  layout.ts        orchestration only

The solver types moved out of layoutRunner.ts, so the runner now depends
on the solver rather than the solver's types living in the module that
happens to run them.

Behaviour-preserving: geometry is byte-identical across a 60-graph corpus
covering all four directions, self-loops, parallel edges, unlabelled
nodes and long labels.
